refactor: reorganize hooks and selectors for better structure and clarity
