+++ src/lib/store/appConfig/appConfig.hooks.ts
@@ -1,25 +1,11 @@
-import { createSelector } from '@reduxjs/toolkit';
-import { AppConfigExtras, RootState } from 'src/lib';
+import { useAppSelector } from '../hooks';
+import { selectApiPath, selectAppConfig, selectLogoPath, selectPartnerMetadata } from './appConfig.selectors';
 
+// Temporary cast to AppConfigExtras until the data is moved to a websocket message
+export const useAppConfig = () => useAppSelector(selectAppConfig);
 
-const appConfigSlice = (state: RootState) => state.appConfig;
+export const useApiPath = () => useAppSelector(selectApiPath);
 
-export const selectAppConfig = createSelector(
-  appConfigSlice,
-  (appConfig) => appConfig.config as AppConfigExtras
-);
+export const useLogoPath = () => useAppSelector(selectLogoPath);
 
-export const selectApiPath = createSelector(
-  appConfigSlice,
-  (appConfig) => appConfig.config.apiPath
-);
-
-export const selectLogoPath = createSelector(
-  appConfigSlice,
-  (appConfig) => appConfig.config.logoPath
-);
-
-export const selectPartnerMetadata = createSelector(
-  appConfigSlice,
-  (appConfig) => appConfig.config.partnerMetadata
-);
+export const usePartnerMetadata = () => useAppSelector(selectPartnerMetadata)
