+++ src/lib/store/devices/devices.hooks.ts
@@ -1,15 +1,22 @@
-import { createSelector } from '@reduxjs/toolkit';
-import { RootState } from 'src/lib';
+import { DeviceState } from 'src/lib';
+import { useAppSelector } from '../hooks';
+import { selectAllDevices, selectDeviceByKey } from './devices.selectors';
 
 
-const devicesConfigSlice = (state: RootState) => state.devices;
+/**
+ * Selector for all devices
+ * @returns Record<string, DeviceState>
+ */
+export const useGetAllDevices = () => {
+  return useAppSelector(selectAllDevices);
+}
 
-export const selectAllDevices = createSelector(
-  devicesConfigSlice,
-  (devices) => devices
-);
+/**
+ * Selector for a single device
+ * @param deviceKey 
+ * @returns DeviceState or undefined
+ */
+export function useGetDevice<T extends DeviceState = DeviceState>(deviceKey: string): T | undefined {
+  return useAppSelector(selectDeviceByKey(deviceKey)) as T | undefined;
+}
 
-export const selectDeviceByKey = (deviceKey: string) => createSelector(
-  devicesConfigSlice,
-  (devices) => devices[deviceKey] ? devices[deviceKey] : undefined
-);
