+++ src/lib/utils/WebsocketProvider.tsx
@@ -6,16 +6,16 @@ import { store, uiActions, useAppConfig } from "../store";
 import { devicesActions } from "../store/devices/devices.slice";
 import { roomsActions } from "../store/rooms/rooms.slice";
 import {
-    useClientId,
-    useRoomKey,
-    useServerIsRunningOnProcessorHardware,
-    useSystemUuid,
-    useUserCode,
-    useWsIsConnected,
-} from "../store/runtimeConfig/runtime.selectors";
+  useClientId,
+  useRoomKey,
+  useServerIsRunningOnProcessorHardware,
+  useSystemUuid,
+  useUserCode,
+  useWsIsConnected,
+} from "../store/runtimeConfig/runtime.hooks";
 import {
-    UserCode,
-    runtimeConfigActions,
+  UserCode,
+  runtimeConfigActions,
 } from "../store/runtimeConfig/runtimeConfig.slice";
 import { Message, RoomData } from "../types";
 import sessionStorageKeys from "../types/classes/session-storage-keys";
